+++ src/main/resources/application.properties
@@ -6,8 +6,10 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/tooliodb
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 #amazon.aws.accesskey=${ACCESS_KEY}
 #amazon.aws.secretkey=${SECRET_KEY}
-#amazon.s3.endpoint=${S3_ENDPOINT}
-#amazon.aws.bucket=${S3_BUCKET}
+amazon.aws.accesskey.s3=${S3_ACCESS_KEY}
+amazon.aws.secretkey.s3=${S3_SECRET_KEY}
+amazon.s3.endpoint=${S3_ENDPOINT}
+amazon.aws.bucket=${S3_BUCKET}
 server.port=5000
 spring.jpa.hibernate.ddl-auto=update
 #spring.jpa.hibernate.ddl-auto=create-drop
