+++ src/main/resources/templates/profile.html
@@ -26,7 +26,7 @@
                 <div class="col">
                     <div class="userTools">
 
-                        <div th:if="${principle">
+                        <div th:if="${principle}">
                             <div th:if="${principle.tools}">
                                 <h3>Tools</h3>
                                 <button type="submit" id="submit" class="btn btn-primary">Add New tool</button>
